refact: move util to mbw_util, for in case namespace trouble
