+++ scripts/mbw/ui_mbw.py
@@ -5,8 +5,8 @@ from modules import sd_models, shared
 from tqdm import tqdm
 
 from scripts.mbw.merge_block_weighted import merge
-from scripts.util.preset_weights import PresetWeights
-from scripts.util.merge_history import MergeHistory
+from scripts.mbw_util.preset_weights import PresetWeights
+from scripts.mbw_util.merge_history import MergeHistory
 
 presetWeights = PresetWeights()
 mergeHistory = MergeHistory()
